+++ segment.cpp
@@ -6,7 +6,7 @@
 
 Segment::Segment() : overflow_(nullptr)
 {
-    table_ = new size_t[kNumOfBuckets * kBucketSize];
+    table_ = new uint32_t[kNumOfBuckets * kBucketSize];
     for (int i = 0; i < kNumOfBuckets * kBucketSize; i++)
     {
         table_[i] = 0;
@@ -22,19 +22,19 @@ Segment::~Segment()
     }
 }
 
-inline size_t calclulate_fingerprint(size_t element)
+inline uint32_t calclulate_fingerprint(uint32_t element)
 {
     // left-most bit set to 1 to indicate that bucket entry is not empty
-    size_t bitmask = ((size_t)-1 >> 1) + 1;
+    uint32_t bitmask = ((uint32_t)-1 >> 1) + 1;
     return element >> (kBucketIndexBitLength + kInitialSegmentIndexBitLength) | bitmask;
 }
 
-bool Segment::insert(size_t element)
+bool Segment::insert(uint32_t element)
 {
     // calculate fingerprint and buckets for this element
-    size_t fingerprint = calclulate_fingerprint(element);
-    size_t i1 = element % kNumOfBuckets;
-    size_t i2 = (i1 ^ fingerprint) % kNumOfBuckets;
+    uint32_t fingerprint = calclulate_fingerprint(element);
+    uint32_t i1 = element % kNumOfBuckets;
+    uint32_t i2 = (i1 ^ fingerprint) % kNumOfBuckets;
 
     // try to put the element in the first bucket
     for (int j = 0; j < kBucketSize; j++)
@@ -57,13 +57,13 @@ bool Segment::insert(size_t element)
     }
 
     // both buckets are full, we need to kick out a random element from one of the buckets
-    size_t i = rand() & 1 == 0 ? i1 : i2;
+    uint32_t i = rand() & 1 == 0 ? i1 : i2;
 
     for (int n = 0; n < kMaxNumKicks; n++)
     {
         // get random entry
         int bucket = rand() % kBucketSize;
-        size_t e = table_[i * kBucketSize + bucket];
+        uint32_t e = table_[i * kBucketSize + bucket];
 
         // put fingerprint in that spot
         table_[i * kBucketSize + bucket] = fingerprint;
@@ -87,16 +87,16 @@ bool Segment::insert(size_t element)
         overflow_ = new Segment();
     }
     // note: new_element doesn't have segment index bits set, but they are not important anymore
-    size_t new_element = fingerprint << (kInitialSegmentIndexBitLength + kBucketIndexBitLength) | i;
+    uint32_t new_element = fingerprint << (kInitialSegmentIndexBitLength + kBucketIndexBitLength) | i;
     return overflow_->insert(new_element);
 }
 
-bool Segment::lookup(size_t element) const
+bool Segment::lookup(uint32_t element) const
 {
     // calculate fingerprint and buckets for this element
-    size_t fingerprint = calclulate_fingerprint(element);
-    size_t i1 = element % kNumOfBuckets;
-    size_t i2 = (i1 ^ fingerprint) % kNumOfBuckets;
+    uint32_t fingerprint = calclulate_fingerprint(element);
+    uint32_t i1 = element % kNumOfBuckets;
+    uint32_t i2 = (i1 ^ fingerprint) % kNumOfBuckets;
 
     // try to find the element in the first bucket
     for (int j = 0; j < kBucketSize; j++)
@@ -126,12 +126,12 @@ bool Segment::lookup(size_t element) const
     return false;
 }
 
-bool Segment::remove(size_t element)
+bool Segment::remove(uint32_t element)
 {
     // calculate fingerprint and buckets for this element
-    size_t fingerprint = calclulate_fingerprint(element);
-    size_t i1 = element % kNumOfBuckets;
-    size_t i2 = (i1 ^ fingerprint) % kNumOfBuckets;
+    uint32_t fingerprint = calclulate_fingerprint(element);
+    uint32_t i1 = element % kNumOfBuckets;
+    uint32_t i2 = (i1 ^ fingerprint) % kNumOfBuckets;
 
     // try to find the element in the first bucket
     for (int j = 0; j < kBucketSize; j++)
@@ -163,7 +163,7 @@ bool Segment::remove(size_t element)
     return false;
 }
 
-void Segment::collect_elements(std::vector<size_t> &elements, size_t segment_index) const
+void Segment::collect_elements(std::vector<uint32_t> &elements, uint32_t segment_index) const
 {
     for (int i = 0; i < kNumOfBuckets; i++)
     {
@@ -171,7 +171,7 @@ void Segment::collect_elements(std::vector<size_t> &elements, size_t segment_ind
         {
             if (table_[i * kBucketSize + j] != 0)
             {
-                size_t element = table_[i * kBucketSize + j] & ~(((size_t)-1 >> 1) + 1); // remove the left-most bit set to 1
+                uint32_t element = table_[i * kBucketSize + j] & ~(((uint32_t)-1 >> 1) + 1); // remove the left-most bit set to 1
                 element <<= (kBucketIndexBitLength + kInitialSegmentIndexBitLength);
                 element |= (segment_index << kBucketIndexBitLength);
                 element |= i;
